+++ docs/scripts/routes/form.js
@@ -1,6 +1,8 @@
 const Textfield = require('../views/textfield');
 const Radio = require('../views/radio');
 const Checkbox = require('../views/checkbox');
+const IconToggle = require('../views/icon-toggle');
+const Switch = require('../views/switch');
 const Select = require('../views/select');
 const Slider = require('../views/slider');
 
@@ -16,6 +18,14 @@ const formRoutes = [{
   path: '/checkbox',
   name: 'checkbox',
   component: Checkbox
+}, {
+  path: '/icon-toggle',
+  name: 'icon-toggle',
+  component: IconToggle
+}, {
+  path: '/switch',
+  name: 'switch',
+  component: Switch
 }, {
   path: '/select',
   name: 'select',